+++ CHANGELOG.md
@@ -7,6 +7,7 @@ ChangeLog
 
 - CLEANUP: cleanup signatures of on/off methods
 - UPDATE: upgrade NPM dependencies
+- IMPROVEMENT: add spooling and runner for error handling
 
 1.3.0 (2026-02-07)
 ------------------